+++ src/Ds/Bundle/DefinitionBundle/DsDefinitionBundle.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Ds\Bundle\DefinitionBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Class DsDefinitionBundle
+ */
+class DsDefinitionBundle extends Bundle
+{
+}