+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This changelog references the relevant changes (new features, changes and bugs) done in 0.4.0 version.
 
   * Added .codeclimate.yml file.
+  * Added definition bundle.
 
 
 # CHANGELOG for 0.3.0